+++ src/app/app.module.ts
@@ -1,49 +1,22 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { HttpClientModule } from '@angular/common/http';
-import { RouterModule } from '@angular/router';
 
-import { AngularFireModule } from 'angularfire2';
-import { AngularFireAuthModule } from 'angularfire2/auth';
-import { AngularFireDatabaseModule } from 'angularfire2/database';
-import { AngularFirestoreModule } from 'angularfire2/firestore';
-import { firebaseConfig } from './firebase.config';
-
-import { FlashMessagesModule } from 'angular2-flash-messages';
-
-import { RoutingModule } from './modules/routing/routing.module';
-import { CollapseDirective } from './directives/collapse.directive';
-import { AuthService } from './services/auth/auth.service';
-import { ElectronService } from './services/electron/electron.service';
-import { ElectronHwidService } from './services/electron-hwid/electron-hwid.service';
-import { IpService } from './services/ip/ip.service';
+import { CoreModule } from '@app/core';
+import { SharedModule } from '@app/shared';
+import { RoutingModule } from '@app/routing';
 import { AppComponent } from './app.component';
-import { NavbarComponent } from './components/navbar/navbar.component';
-import { FooterComponent } from './components/footer/footer.component';
 
 @NgModule({
   declarations: [
     AppComponent,
-    CollapseDirective,
-    NavbarComponent,
-    FooterComponent
   ],
   imports: [
     BrowserModule,
-    HttpClientModule,
+    CoreModule,
+    SharedModule,
     RoutingModule,
-    AngularFireModule.initializeApp(firebaseConfig),
-    AngularFireAuthModule,
-    AngularFireDatabaseModule,
-    AngularFirestoreModule.enablePersistence(),
-    FlashMessagesModule
   ],
-  providers: [
-    AuthService,
-    ElectronService,
-    ElectronHwidService,
-    IpService
-  ],
+  providers: [],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
