+++ src/app/core/firebase.config.example.ts
