+++ src/app/core/index.ts
@@ -1,0 +1,11 @@
+export * from './core.module';
+
+export * from './guards/auth/auth.guard';
+export * from './guards/guest/guest.guard';
+
+export * from './models/role';
+export * from './models/user';
+
+export * from './services/auth/auth.service';
+export * from './services/electron/electron.service';
+export * from './services/ip/ip.service';
