+++ src/app/core/services/electron-hwid/electron-hwid.service.ts
