+++ src/app/core/services/electron/electron.service.spec.ts
