+++ src/app/material/index.ts
@@ -1,0 +1,1 @@
+export * from './material.module';
