+++ src/app/routing/index.ts
@@ -1,0 +1,5 @@
+export * from './routing.module';
+
+export * from './components/home/home.component';
+export * from './components/login-form/login-form.component';
+export * from './components/register-form/register-form.component';
