+++ src/app/shared/components/footer/footer.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 
-import { AuthService } from '../../services/auth/auth.service';
-import { Role } from './../../models/role';
+import { AuthService, Role } from '@app/core';
 
 @Component({
   selector: 'app-footer',
