+++ src/app/shared/components/navbar/navbar.component.scss
