+++ src/app/shared/index.ts
@@ -1,0 +1,9 @@
+export * from './shared.module';
+
+export * from './components/footer/footer.component';
+export * from './components/navbar/navbar.component';
+export * from './components/page/page.component';
+
+export * from './directives/collapse/collapse.directive';
+
+export * from './pipes/circular-json/circular-json.pipe';
