+++ src/app/shared/pipes/circular-json/circular-json.pipe.ts
